Python Version Nearly Complete
Just formatting and optimizing left to go.
